api/requirements.txt
fastapi==0.115.11
instagrapi==2.1.3
pydantic==2.10.1
uvicorn==0.34.0
pillow==11.1.0
python-dotenv==1.0.1
